ip/requirements.txt
vunit_hdl>=5.0.0.dev5

# packages for TerosHDL
teroshdl
cocotb
yowasp-yosys
edalize
vsg
